+++ src/hooks/useAnnotationEvents.ts
@@ -1,4 +1,4 @@
-import { AnnotationEvent, TimeRange } from '@grafana/data';
+import { AnnotationEvent, DataFrame, PanelData, TimeRange } from '@grafana/data';
 import { getBackendSrv } from '@grafana/runtime';
 import dayjs from 'dayjs';
 import { useEffect, useMemo, useState } from 'react';
@@ -6,11 +6,11 @@ import { useEffect, useMemo, useState } from 'react';
 import { AnnotationsType, CalendarEvent, CalendarOptions } from '../types';
 
 /**
- * Get Annotations
+ * Get Api Annotations
  * @param timeRange
  * @param options
  */
-const useAnnotations = (timeRange: TimeRange, options: CalendarOptions) => {
+const useApiAnnotations = (timeRange: TimeRange, options: CalendarOptions) => {
   const [annotations, setAnnotations] = useState<AnnotationEvent[]>([]);
 
   useEffect(() => {
@@ -37,8 +37,82 @@ const useAnnotations = (timeRange: TimeRange, options: CalendarOptions) => {
 
     getBackendSrv()
       .get<AnnotationEvent[] | null>('/api/annotations', params)
-      .then((res) => setAnnotations(Array.isArray(res) ? res : []));
+      .then((res) => {
+        let apiEvents: AnnotationEvent[] = [];
+
+        if (res && Array.isArray(res)) {
+          apiEvents = res.map((event) => ({
+            ...event,
+            title: event.title || event.text,
+            text: !event.title ? '' : event.text,
+          }));
+        }
+
+        setAnnotations(apiEvents);
+      });
   }, [timeRange, options.annotationsLimit, options.annotationsType]);
+
+  return annotations;
+};
+
+/**
+ * Get Dashboard Annotations
+ * @param timeRange
+ * @param data
+ */
+const useDashboardAnnotations = (timeRange: TimeRange, dashboardAnnotations?: DataFrame[]) => {
+  const [annotations, setAnnotations] = useState<AnnotationEvent[]>([]);
+
+  useEffect(() => {
+    if (!!dashboardAnnotations?.length) {
+      /**
+       * Get dashboard annotations
+       */
+      const annotations = dashboardAnnotations.flatMap((annotation) => {
+        /**
+         * Get necessary fields
+         */
+        const title = annotation.fields.find((field) => field.name === 'title');
+        const tags = annotation.fields.find((field) => field.name === 'tags');
+        const color = annotation.fields.find((field) => field.name === 'color');
+        const time = annotation.fields.find((field) => field.name === 'time');
+        const timeEnd = annotation.fields.find((field) => field.name === 'timeEnd');
+
+        /**
+         * Text use for description
+         */
+        const text = annotation.fields.find((field) => field.name === 'text');
+
+        /**
+         * Return annotations
+         */
+        return Array.from(Array(annotation.length)).map((event, index) => {
+          return {
+            title: title?.values[index] || '',
+            tags: tags?.values[index] || [],
+            color: color?.values[index] || '',
+            time: time?.values[index] || undefined,
+            text: text?.values[index] || '',
+            timeEnd: timeEnd?.values[index] || undefined,
+          };
+        });
+      });
+
+      /**
+       * Filter annotations by time range
+       * Define start and end dates
+       */
+      const startDate = timeRange.from.valueOf();
+      const endDate = timeRange.to.valueOf();
+
+      const filteredAnnotations = annotations.filter((item) => {
+        const itemDate = dayjs(item.time);
+        return itemDate.isAfter(startDate) && itemDate.isBefore(endDate);
+      });
+
+      setAnnotations(filteredAnnotations);
+    }
+  }, [dashboardAnnotations, timeRange]);
 
   return annotations;
 };
@@ -48,17 +122,19 @@ const useAnnotations = (timeRange: TimeRange, options: CalendarOptions) => {
  * @param timeRange
  * @param options
  */
-export const useAnnotationEvents = (timeRange: TimeRange, options: CalendarOptions) => {
-  const annotations = useAnnotations(timeRange, options);
+export const useAnnotationEvents = (timeRange: TimeRange, options: CalendarOptions, data: PanelData) => {
+  const apiAnnotations = useApiAnnotations(timeRange, options);
+  const dashboardAnnotations = useDashboardAnnotations(timeRange, data.annotations);
 
   return useMemo(() => {
-    return annotations.map<CalendarEvent>((annotation) => ({
-      text: annotation.text ?? '',
+    return [...apiAnnotations, ...dashboardAnnotations].map<CalendarEvent>((annotation) => ({
+      text: annotation.title ?? '',
       start: dayjs(annotation.time),
       end: annotation.timeEnd ? dayjs(annotation.timeEnd) : undefined,
       open: false,
       labels: annotation.tags || [],
+      description: annotation.text ?? '',
       color: annotation.color || '',
     }));
-  }, [annotations]);
+  }, [apiAnnotations, dashboardAnnotations]);
 };