+++ src/module.ts
@@ -303,4 +303,5 @@ export const plugin = new PanelPlugin<CalendarOptions>(CalendarPanel)
         },
         category: [t('panelOptions.annotations.label')],
       });
-  });
+  })
+  .setDataSupport({ annotations: true });