+++ CHANGELOG.md
@@ -8,9 +8,11 @@
 - Added Formatting for event description (#182)
 - Updated Details drawer layout (#192)
 - Added support for additional Time Range sources (#179)
+- Update event format (#188)
 - Updated Time Range in Agenda view (#196)
 - Updated mouse cursor for Agenda view (#200)
 - Updated dependencies (#200)
+- Added display dashboard annotations (#194)
 
 ## 3.1.0 (2024-05-19)